Sorry for late commit!
I did indeed have this work done in time, as shown on my milestone 1 submission being on time and having everything here

+++ C292 Midterm/Assets/RuntimeData.cs
@@ -6,6 +6,5 @@ using UnityEngine;
 
 public class RuntimeData : ScriptableObject 
 {
-    public float CurrentLevel = 1;
-    
+    public int CurrentLevel = 1;
 }
